+++ pkg/JuliaInterface/gap/JuliaInterface.gd
@@ -184,6 +184,26 @@ BindGlobal( "TheTypeOfJuliaModules", NewType( TheFamilyOfJuliaModules, IsJuliaMo
 #! @EndExampleSession
 #DeclareGlobalFunction( "JuliaEvalString" );
 
+#! @Arguments T, parameters
+#! @Description
+#!  creates the parametrized &Julia; type
+#!  <A>T</A><C>{para1, para2, ...}</C>,
+#!  where <A>T</A> is a &Julia; type and <C>para1</C>, <C>para2</C>, ...
+#!  describe its intended type parameters.
+#!  Each of <C>para1</C>, <C>para2</C>, ... must be either a &Julia; type
+#!  or a list of the form <C>[ T2, parameters2 ]</C>,
+#!  where the latter means the type <C>JuliaType( T2, parameters2 )</C>.
+#! @BeginExampleSession
+#! gap> JuliaType( Julia.Vector, [ Julia.Int ] );
+#! <Julia: Vector{Int64}>
+#! gap> JuliaType( Julia.Tuple, [ Julia.Int, Julia.Int ] );
+#! <Julia: Tuple{Int64, Int64}>
+#! gap> JuliaType( Julia.Tuple, [ [ Julia.Vector, [ Julia.Int ] ],
+#! >                              [ Julia.Matrix, [ Julia.Int ] ] ] );
+#! <Julia: Tuple{Vector{Int64}, Matrix{Int64}}>
+#! @EndExampleSession
+DeclareGlobalFunction( "JuliaType" );
+
 #! @Arguments filename[, module_name]
 #! @Returns nothing.
 #! @Description