+++ src/utils.jl
@@ -111,6 +111,16 @@ function kwarg_wrapper(func, args::Vector{T1}, kwargs::Dict{Symbol,T2}) where {T
     return func(args...; [k => kwargs[k] for k in keys(kwargs)]...)
 end
 
+
+## helper for creating parametrized Julia types in GAP
+
+create_type(T::Type) = T
+
+function create_type(T::Type, paras::Vector)
+  return T{[((x isa Type) ? x : create_type(x...)) for x in paras]...}
+end
+
+
 ## convenience function
 
 function Display(x::GapObj)